Rename all WebSocket code into one directory.
