+++ Builds/VisualStudio2013/RippleD.vcxproj
@@ -1959,20 +1959,6 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='debug|x64'">True</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='release|x64'">True</ExcludedFromBuild>
     </ClCompile>
-    <ClInclude Include="..\..\src\ripple\app\websocket\WSConnection.h">
-    </ClInclude>
-    <ClCompile Include="..\..\src\ripple\app\websocket\WSDoor.cpp">
-      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='debug|x64'">True</ExcludedFromBuild>
-      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='release|x64'">True</ExcludedFromBuild>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\app\websocket\WSDoor.h">
-    </ClInclude>
-    <ClCompile Include="..\..\src\ripple\app\websocket\WSServerHandler.cpp">
-      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='debug|x64'">True</ExcludedFromBuild>
-      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='release|x64'">True</ExcludedFromBuild>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\app\websocket\WSServerHandler.h">
-    </ClInclude>
     <ClInclude Include="..\..\src\ripple\basics\base_uint.h">
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\basics\BasicConfig.h">
@@ -3476,8 +3462,6 @@
     </ClCompile>
     <ClCompile Include="..\..\src\ripple\unity\websocket.cpp">
     </ClCompile>
-    <ClInclude Include="..\..\src\ripple\unity\websocket.h">
-    </ClInclude>
     <ClInclude Include="..\..\src\ripple\validators\Connection.h">
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\validators\impl\ConnectionImp.cpp">
@@ -3509,14 +3493,20 @@
     <ClCompile Include="..\..\src\ripple\validators\tests\Validators.test.cpp">
       <ExcludedFromBuild>True</ExcludedFromBuild>
     </ClCompile>
-    <ClCompile Include="..\..\src\ripple\websocket\autosocket\AutoSocket.cpp">
-      <ExcludedFromBuild>True</ExcludedFromBuild>
-    </ClCompile>
-    <ClInclude Include="..\..\src\ripple\websocket\autosocket\AutoSocket.h">
+    <ClInclude Include="..\..\src\ripple\websocket\AutoSocket.h">
     </ClInclude>
-    <ClCompile Include="..\..\src\ripple\websocket\autosocket\LogWebsockets.cpp">
+    <ClInclude Include="..\..\src\ripple\websocket\Connection.h">
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\websocket\Handler.h">
+    </ClInclude>
+    <ClCompile Include="..\..\src\ripple\websocket\LogWebsockets.cpp">
       <ExcludedFromBuild>True</ExcludedFromBuild>
     </ClCompile>
+    <ClCompile Include="..\..\src\ripple\websocket\MakeServer.cpp">
+      <ExcludedFromBuild>True</ExcludedFromBuild>
+    </ClCompile>
+    <ClInclude Include="..\..\src\ripple\websocket\MakeServer.h">
+    </ClInclude>
     <ClCompile Include="..\..\src\rocksdb2\db\builder.cc">
       <ExcludedFromBuild>True</ExcludedFromBuild>
     </ClCompile>
