+++ src/ripple/unity/app.cpp
@@ -18,16 +18,12 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/rpc/Manager.h>
-#include <ripple/unity/websocket.h>
 #include <ripple/app/impl/BasicApp.cpp>
-#include <ripple/app/main/CollectorManager.cpp>
-#include <ripple/app/main/NodeStoreScheduler.cpp>
-#include <ripple/app/websocket/WSServerHandler.cpp>
-#include <ripple/app/websocket/WSConnection.cpp>
-#include <ripple/app/websocket/WSDoor.cpp>
-#include <ripple/app/node/SqliteFactory.cpp>
 #include <ripple/app/main/Application.cpp>
+#include <ripple/app/main/CollectorManager.cpp>
 #include <ripple/app/main/Main.cpp>
+#include <ripple/app/main/NodeStoreScheduler.cpp>
+#include <ripple/app/node/SqliteFactory.cpp>
 #include <ripple/resource/Manager.h>
+#include <ripple/rpc/Manager.h>
 #include <beast/module/core/time/Time.h>
