+++ src/ripple/websocket/AutoSocket.h
