+++ src/websocketpp_02/src/sockets/autotls.hpp
@@ -28,7 +28,7 @@
 #ifndef WEBSOCKETPP_SOCKET_AUTOTLS_HPP
 #define WEBSOCKETPP_SOCKET_AUTOTLS_HPP
 
-#include <ripple/websocket/autosocket/AutoSocket.h>
+#include <ripple/websocket/AutoSocket.h>
 #include <beast/asio/placeholders.h>
 #include <functional>
